% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_jsd.R
\name{calculate_jsd}
\alias{calculate_jsd}
\title{Calculate Jensen-Shannon divergence}
\usage{
calculate_jsd(intensities1, intensities2)
}
\arguments{
\item{intensities2}{}
}
\description{
Calculate Jensen-Shannon divergence
}
