% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_intensities_df.R
\name{create_intensities_df}
\alias{create_intensities_df}
\title{Create intensities df}
\usage{
create_intensities_df(df_intensities_ot, df_intensities_tof)
}
\arguments{
\item{df_intensities_tof}{}
}
\description{
Create intensities df
}
