% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\docType{data}
\name{dependencies}
\alias{dependencies}
\title{Dependencies to be loaded from CRAN, Bioconductor and Github}
\format{
An object of class \code{list} of length 3.
}
\usage{
dependencies
}
\description{
Dependencies to be loaded from CRAN, Bioconductor and Github
}
\keyword{datasets}
