% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_figure.R
\name{export_figure}
\alias{export_figure}
\title{Export figure}
\usage{
export_figure(figure, filename, extension = EXTENSION, height = 7, width = 7)
}
\arguments{
\item{width}{}
}
\description{
Export figure
}
