% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fragments.R
\name{filter_fragments}
\alias{filter_fragments}
\title{Filter fragments}
\usage{
filter_fragments(
  fragments,
  diffs,
  diff_min = -4.5,
  diffs_to_keep = list(c(-4.03, -3.99), c(-2.03, -1.99)),
  fragments_to_exclude = list(c(143.7, 143.74), c(173.5, 173.54)),
  invert_selection = FALSE
)
}
\arguments{
\item{invert_selection}{}
}
\description{
Filter fragments
}
