% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_standards.R
\name{import_standards}
\alias{import_standards}
\title{Import dirs}
\usage{
import_standards(
  file,
  columns,
  score_min = 0.7,
  remove_above = TRUE,
  diff_min = -4.5,
  diffs_to_keep = list(c(-4.03, -3.99), c(-2.03, -1.99)),
  fragments_to_exclude = list(c(143.7, 143.74), c(173.5, 173.54))
)
}
\arguments{
\item{fragments_to_exclude}{}
}
\description{
Import dirs
}
