% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_unknowns.R
\name{import_unknowns}
\alias{import_unknowns}
\title{Import dirs}
\usage{
import_unknowns(
  file,
  columns,
  remove_above = TRUE,
  diff_min = -4.5,
  diffs_to_keep = list(c(-4.03, -3.99), c(-2.03, -1.99)),
  fragments_to_exclude = list(c(143.7, 143.74), c(173.5, 173.54)),
  fragmented_intensity_min = 0.5
)
}
\arguments{
\item{fragmented_intensity_min}{}
}
\description{
Import dirs
}
