% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dependencies.R
\name{install_dependencies_2}
\alias{install_dependencies_2}
\title{Install dependencies 2}
\usage{
install_dependencies_2(dependencies)
}
\arguments{
\item{dependencies}{}
}
\description{
Install dependencies 2
}
