% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionTypeAnalysis-package.R
\docType{package}
\name{ionTypeAnalysis-package}
\alias{ionTypeAnalysis}
\alias{ionTypeAnalysis-package}
\title{ionTypeAnalysis: Ion Type Analysis}
\description{
This package provides the infrastructure to perform Ion Type Analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zamboni-lab/ion-type-analysis}
  \item Report bugs at \url{https://github.com/zamboni-lab/ion-type-analysis/issues}
}

}
\author{
\strong{Maintainer}: Adriano Rutz \email{rutz@imsb.biol.ethz.ch} (\href{https://orcid.org/0000-0003-0443-9902}{ORCID})

}
\keyword{internal}
