% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_compound_classes.R
\name{load_compound_classes}
\alias{load_compound_classes}
\title{Load compound classes}
\usage{
load_compound_classes(
  files = c("~/../../Volumes/T7_Shield/enamdisc_cleaned.tsv",
    "~/../../Volumes/T7_Shield/enammol_cleaned.tsv",
    "~/../../Volumes/T7_Shield/mcebio_cleaned.tsv",
    "~/../../Volumes/T7_Shield/mcedrug_cleaned.tsv",
    "~/../../Volumes/T7_Shield/mcescaf_cleaned.tsv",
    "~/../../Volumes/T7_Shield/nihnp_cleaned.tsv",
    "~/../../Volumes/T7_Shield/otavapep_cleaned.tsv")
)
}
\arguments{
\item{files}{}
}
\description{
Load compound classes
}
