% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bins.R
\name{plot_bins}
\alias{plot_bins}
\title{Import dirs}
\usage{
plot_bins(
  mzs,
  bin_size = 1,
  min_mz = -180.5,
  max_mz = 50.5,
  threshold = 10,
  y_min = 0,
  y_max = 70,
  mutate = TRUE,
  title = NULL
)
}
\arguments{
\item{title}{}
}
\description{
Import dirs
}
