% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Plot distributions}
\usage{
plot_distributions(
  df,
  facet,
  group,
  value,
  axis_label = "",
  title = "",
  subtitle = "",
  caption = "",
  reorder = NULL,
  plot_histograms = FALSE,
  normalize_xlim = TRUE
)
}
\arguments{
\item{normalize_xlim}{}
}
\description{
Plot distributions
}
