% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions_relative_intensities.R
\name{plot_distributions_relative_intensities}
\alias{plot_distributions_relative_intensities}
\title{Plot distributions relative intensities}
\usage{
plot_distributions_relative_intensities(
  df,
  type_1 = "more than 2 -OH",
  type_2 = "2 -OH or less",
  pal = c("#DC0000FF", "#E64B35FF", "#F39B7FFF")
)
}
\arguments{
\item{df}{Df}

\item{type_1}{Type 1}

\item{type_2}{Type 2}

\item{pal}{Pal}
}
\description{
Plot distributions relative intensities
}
