% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainplots_by_group.R
\name{rainplots_by_group}
\alias{rainplots_by_group}
\title{Title TODO}
\usage{
rainplots_by_group(
  df,
  group,
  value,
  axis_label = "",
  title = "",
  subtitle = "",
  caption = ""
)
}
\arguments{
\item{caption}{}
}
\description{
Title TODO
}
