% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_table.R
\name{reformat_table}
\alias{reformat_table}
\title{Reformat table}
\usage{
reformat_table(
  table,
  file,
  remove_above = TRUE,
  diff_min = -4.5,
  diffs_to_keep = list(c(-4.03, -3.99), c(-2.03, -1.99)),
  fragments_to_exclude = list(c(143.7, 143.74), c(173.5, 173.54))
)
}
\arguments{
\item{fragments_to_exclude}{}
}
\description{
Reformat table
}
