% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_highlight_regions.R
\name{venn_highlight_regions}
\alias{venn_highlight_regions}
\title{Venn highlight regions}
\usage{
venn_highlight_regions(df, regions, color = "yellow")
}
\arguments{
\item{color}{}
}
\description{
Venn highlight regions
}
