% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_plot_highlighted.R
\name{venn_plot_highlighted}
\alias{venn_plot_highlighted}
\title{Venn plot highlighted}
\usage{
venn_plot_highlighted(df_venn, venn_simple)
}
\arguments{
\item{venn_simple}{}
}
\description{
Venn plot highlighted
}
