% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions_relative_intensities.R
\name{plot_distributions_relative_intensities}
\alias{plot_distributions_relative_intensities}
\title{Plot distributions relative intensities}
\usage{
plot_distributions_relative_intensities(
  df,
  type_1 = "more than 2 -OH",
  type_2 = "2 -OH or less"
)
}
\arguments{
\item{df}{}
}
\description{
Plot distributions relative intensities
}
