% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_intervals.R
\name{select_intervals}
\alias{select_intervals}
\title{Find best combination of windows}
\usage{
select_intervals(df)
}
\arguments{
\item{df}{}
}
\value{
A list containing the best filtered df and best summed height
}
\description{
Find best combination of windows
}
